+++ pkg/store/proxy_merge_test.go
@@ -5,10 +5,15 @@ package store
 
 import (
 	"fmt"
+	"io"
 	"sync"
 	"testing"
 
 	"github.com/efficientgo/core/testutil"
+	"github.com/opentracing/opentracing-go"
+	otlog "github.com/opentracing/opentracing-go/log"
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/promauto"
 	"github.com/prometheus/prometheus/model/labels"
 
 	"github.com/thanos-io/thanos/pkg/errors"
@@ -241,6 +246,171 @@ type nopClientSendCloser struct {
 
 func (c nopClientSendCloser) CloseSend() error { return nil }
 
+// TestProxyResponseTreeSortWithBatchResponses verifies that batch responses are
+// properly unpacked into individual series before being merged by the loser tree.
+// Without proper unpacking, the loser tree would receive batch responses and fail
+// to merge series correctly.
+func TestProxyResponseTreeSortWithBatchResponses(t *testing.T) {
+	t.Parallel()
+
+	// Create series that will be sent in batches from two stores.
+	// Store 1 sends batch with series a=1, a=3
+	// Store 2 sends batch with series a=2, a=4
+	// After unpacking and merging, we expect: a=1, a=2, a=3, a=4
+	series1 := storeSeriesResponse(t, labelsFromStrings("a", "1")).GetSeries()
+	series2 := storeSeriesResponse(t, labelsFromStrings("a", "2")).GetSeries()
+	series3 := storeSeriesResponse(t, labelsFromStrings("a", "3")).GetSeries()
+	series4 := storeSeriesResponse(t, labelsFromStrings("a", "4")).GetSeries()
+
+	mockClient1 := &mockBatchSeriesClient{
+		responses: []*storepb.SeriesResponse{
+			storepb.NewBatchResponse([]*storepb.Series{series1, series3}),
+		},
+	}
+	mockClient2 := &mockBatchSeriesClient{
+		responses: []*storepb.SeriesResponse{
+			storepb.NewBatchResponse([]*storepb.Series{series2, series4}),
+		},
+	}
+
+	// Create eagerRespSets which should unpack the batches
+	var shardInfo *storepb.ShardInfo
+	respSet1 := newEagerRespSet(
+		noopSpan{},
+		0,
+		"store1",
+		nil,
+		func() {},
+		mockClient1,
+		shardInfo.Matcher(nil),
+		false,
+		promauto.NewCounter(prometheus.CounterOpts{Name: fmt.Sprintf("%s-1", t.Name())}),
+		nil,
+	)
+	respSet2 := newEagerRespSet(
+		noopSpan{},
+		0,
+		"store2",
+		nil,
+		func() {},
+		mockClient2,
+		shardInfo.Matcher(nil),
+		false,
+		promauto.NewCounter(prometheus.CounterOpts{Name: fmt.Sprintf("%s-2", t.Name())}),
+		nil,
+	)
+
+	h := NewProxyResponseLoserTree(respSet1, respSet2)
+	var got []*storepb.SeriesResponse
+	for h.Next() {
+		got = append(got, h.At())
+	}
+
+	exp := []*storepb.SeriesResponse{
+		storeSeriesResponse(t, labelsFromStrings("a", "1")),
+		storeSeriesResponse(t, labelsFromStrings("a", "2")),
+		storeSeriesResponse(t, labelsFromStrings("a", "3")),
+		storeSeriesResponse(t, labelsFromStrings("a", "4")),
+	}
+	testutil.Equals(t, exp, got)
+}
+
+type mockBatchSeriesClient struct {
+	storepb.Store_SeriesClient
+	responses []*storepb.SeriesResponse
+	i         int
+}
+
+func (c *mockBatchSeriesClient) Recv() (*storepb.SeriesResponse, error) {
+	if c.i >= len(c.responses) {
+		return nil, io.EOF
+	}
+	resp := c.responses[c.i]
+	c.i++
+	return resp, nil
+}
+
+func (c *mockBatchSeriesClient) CloseSend() error { return nil }
+
+// noopSpan implements opentracing.Span for testing.
+type noopSpan struct{}
+
+func (noopSpan) Finish()                                        {}
+func (noopSpan) SetTag(string, any) opentracing.Span            { return noopSpan{} }
+func (noopSpan) Context() opentracing.SpanContext               { return nil }
+func (noopSpan) SetOperationName(string) opentracing.Span       { return noopSpan{} }
+func (noopSpan) Tracer() opentracing.Tracer                     { return nil }
+func (noopSpan) SetBaggageItem(string, string) opentracing.Span { return noopSpan{} }
+func (noopSpan) BaggageItem(string) string                      { return "" }
+func (noopSpan) LogKV(...any)                                   {}
+func (noopSpan) LogFields(...otlog.Field)                       {}
+func (noopSpan) Log(opentracing.LogData)                        {}
+func (noopSpan) FinishWithOptions(opentracing.FinishOptions)    {}
+func (noopSpan) LogEvent(string)                                {}
+func (noopSpan) LogEventWithPayload(string, interface{})        {}
+
+// TestLazyRespSetUnpacksBatchResponses verifies that lazyRespSet properly unpacks
+// batch responses into individual series before being merged by the loser tree.
+func TestLazyRespSetUnpacksBatchResponses(t *testing.T) {
+	t.Parallel()
+
+	series1 := storeSeriesResponse(t, labelsFromStrings("a", "1")).GetSeries()
+	series2 := storeSeriesResponse(t, labelsFromStrings("a", "2")).GetSeries()
+	series3 := storeSeriesResponse(t, labelsFromStrings("a", "3")).GetSeries()
+	series4 := storeSeriesResponse(t, labelsFromStrings("a", "4")).GetSeries()
+
+	mockClient1 := &mockBatchSeriesClient{
+		responses: []*storepb.SeriesResponse{
+			storepb.NewBatchResponse([]*storepb.Series{series1, series3}),
+		},
+	}
+	mockClient2 := &mockBatchSeriesClient{
+		responses: []*storepb.SeriesResponse{
+			storepb.NewBatchResponse([]*storepb.Series{series2, series4}),
+		},
+	}
+
+	var shardInfo *storepb.ShardInfo
+	respSet1 := newLazyRespSet(
+		noopSpan{},
+		0,
+		"store1",
+		nil,
+		func() {},
+		mockClient1,
+		shardInfo.Matcher(nil),
+		false,
+		promauto.NewCounter(prometheus.CounterOpts{Name: fmt.Sprintf("%s-1", t.Name())}),
+		10,
+	)
+	respSet2 := newLazyRespSet(
+		noopSpan{},
+		0,
+		"store2",
+		nil,
+		func() {},
+		mockClient2,
+		shardInfo.Matcher(nil),
+		false,
+		promauto.NewCounter(prometheus.CounterOpts{Name: fmt.Sprintf("%s-2", t.Name())}),
+		10,
+	)
+
+	h := NewProxyResponseLoserTree(respSet1, respSet2)
+	var got []*storepb.SeriesResponse
+	for h.Next() {
+		got = append(got, h.At())
+	}
+
+	exp := []*storepb.SeriesResponse{
+		storeSeriesResponse(t, labelsFromStrings("a", "1")),
+		storeSeriesResponse(t, labelsFromStrings("a", "2")),
+		storeSeriesResponse(t, labelsFromStrings("a", "3")),
+		storeSeriesResponse(t, labelsFromStrings("a", "4")),
+	}
+	testutil.Equals(t, exp, got)
+}
+
 func TestSortWithoutLabels(t *testing.T) {
 	t.Parallel()
 
@@ -348,7 +518,6 @@ func BenchmarkSortWithoutLabels(b *testing.B) {
 	}
 
 	b.ReportAllocs()
-
 	for b.Loop() {
 		b.StopTimer()
 		for i := range int(1e4) {